refactoring : moved the method that create the Toggle FullScreen Button to the parent class, so that it can be reused by other subclasses
