+++ gazeplay/src/main/java/net/gazeplay/GraphicalContext.java
@@ -1,13 +1,22 @@
 package net.gazeplay;
 
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
+import javafx.event.Event;
+import javafx.event.EventHandler;
 import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
 import lombok.Data;
 import lombok.Getter;
+import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.Configuration;
 import net.gazeplay.commons.configuration.ConfigurationBuilder;
@@ -50,6 +59,45 @@ public abstract class GraphicalContext<T> {
         getChildren().clear();
 
         log.info("Nodes not removed: {}", getChildren().size());
+    }
+
+    public Button createToggleFullScreenButtonInGameScreen(@NonNull GazePlay gazePlay) {
+
+        EventHandler<Event> eventHandler = new EventHandler<javafx.event.Event>() {
+            @Override
+            public void handle(javafx.event.Event e) {
+                gazePlay.toggleFullScreen();
+            }
+        };
+
+        Button button = new Button();
+        configureFullScreenToggleButton(gazePlay.isFullScreen(), button);
+
+        gazePlay.getFullScreenProperty().addListener(new ChangeListener<Boolean>() {
+            @Override
+            public void changed(ObservableValue<? extends Boolean> observable, Boolean wasFullScreen,
+                    Boolean isFullScreen) {
+                configureFullScreenToggleButton(isFullScreen, button);
+            }
+        });
+
+        button.addEventHandler(MouseEvent.MOUSE_CLICKED, eventHandler);
+
+        return button;
+    }
+
+    private void configureFullScreenToggleButton(Boolean isFullScreen, Button button) {
+        final Image buttonGraphics;
+        final String label;
+        if (isFullScreen) {
+            buttonGraphics = new Image("data/common/images/fullscreen-exit.png");
+            label = "Exit FullScreen";
+        } else {
+            buttonGraphics = new Image("data/common/images/fullscreen-enter.png");
+            label = "Enter FullScreen";
+        }
+        button.setGraphic(new ImageView(buttonGraphics));
+        button.setText(label);
     }
 
     public void onGameStarted() {
